feat(nx-plugin): add e2e utils - create nx e2e environment
