+++ packages/nx-plugin/src/schematics/plugin/files/plugin/collection.json__tmpl__
@@ -1,0 +1,12 @@
+{
+  "$schema": "<%= offsetFromRoot %>node_modules/@angular-devkit/schematics/collection-schema.json",
+  "name": "<%= name %>",
+  "version": "0.0.1",
+  "schematics": {
+    "<%= propertyName %>": {
+      "factory": "./src/schematics/<%= fileName %>/schematic",
+      "schema": "./src/schematics/<%= fileName %>/schema.json",
+      "description": "<%= name %> schematic"
+    }
+  }
+}
