+++ packages/nx-plugin/src/schematics/plugin/files/plugin/package.json__tmpl__
@@ -2,6 +2,6 @@
     "name": "<%= npmPackageName %>",
     "version": "0.0.1",
     "main": "src/index.js",
-    "schematics": "src/collection.json",
-    "builders": "src/builders.json",
+    "schematics": "./collection.json",
+    "builders": "./builders.json",
 }
