--- packages/nx-plugin/src/schematics/plugin/files/plugin/src/builders.json__tmpl__
@@ -1,10 +1,0 @@
-{
-  "$schema": "<%= offsetFromRoot %>../node_modules/@angular-devkit/architect/src/builders-schema.json",
-  "builders": {
-    "build": {
-      "implementation": "./builders/<%= fileName %>/builder",
-      "schema": "./builders/<%= fileName %>/schema.json",
-      "description": "<%= name %> builder"
-    }
-  }
-}
