--- packages/nx-plugin/src/schematics/plugin/files/plugin/src/collection.json__tmpl__
@@ -1,12 +1,0 @@
-{
-  "$schema": "<%= offsetFromRoot %>../node_modules/@angular-devkit/schematics/collection-schema.json",
-  "name": "<%= name %>",
-  "version": "0.0.1",
-  "schematics": {
-    "<%= propertyName %>": {
-      "factory": "./schematics/<%= fileName %>/schematic",
-      "schema": "./schematics/<%= fileName %>/schema.json",
-      "description": "<%= name %> schematic"
-    }
-  }
-}
